+++ app/javascript/vue/tasks/collection_objects/filter/components/filters/loan.vue
@@ -20,21 +20,43 @@
       </li>
       <li>
         <label>
-          <input 
+          <input
             v-model="loans.never_loaned"
             type="checkbox">
           Never loaned
         </label>
       </li>
     </ul>
+    <autocomplete
+      class="margin-medium-top"
+      url="/loans/autocomplete"
+      param="term"
+      clear-after
+      placeholder="Search a loan..."
+      label="label_html"
+      @getItem="addLoan($event.id)"
+    />
+    <display-list
+      :list="loanList"
+      label="object_tag"
+      :delete-warning="false"
+      @deleteIndex="removeLoan"/>
   </div>
 </template>
 
 <script>
 
 import { URLParamsToJSON } from 'helpers/url/parse.js'
+import { Loan } from 'routes/endpoints'
+import Autocomplete from 'components/ui/Autocomplete.vue'
+import DisplayList from 'components/displayList.vue'
 
 export default {
+  components: {
+    Autocomplete,
+    DisplayList
+  },
+
   props: {
     modelValue: {
       type: Object,
@@ -43,6 +65,12 @@ export default {
   },
 
   emits: ['update:modelValue'],
+
+  data () {
+    return {
+      loanList: []
+    }
+  },
 
   computed: {
     loans: {
@@ -54,11 +82,36 @@ export default {
       }
     }
   },
-  mounted () {
+
+  watch: {
+    loanList: {
+      handler (newVal) {
+        this.loans.loan_id = newVal.map(item => item.id)
+      },
+      deep: true
+    }
+  },
+
+  created () {
     const urlParams = URLParamsToJSON(location.href)
+    const loanIds = urlParams.loan_id || []
+
     this.loans.on_loan = urlParams.on_loan
     this.loans.loaned = urlParams.loaned
     this.loans.never_loaned = urlParams.never_loaned
+    loanIds.forEach(id => this.addLoan(id))
+  },
+
+  methods: {
+    addLoan (id) {
+      Loan.find(id).then(({ body }) => {
+        this.loanList.push(body)
+      })
+    },
+
+    removeLoan (index) {
+      this.loanList.splice(index, 1)
+    }
   }
 }
 </script>