+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 - Support for DwC terms on depth [#2628]
+- Filter by individual Loan [#2632]
 
 ### Fixed
 - Determiners are not saved after lock the list in comprehensive form [#2626]